backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.2.1
dnspython==2.7.0
eventlet==0.40.3
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fsspec==2025.9.0
google-auth==2.40.3
google-genai==1.32.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instaloader==4.14.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.22.1
packaging==25.0
pillow==11.3.0
proglog==0.1.12
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyreadline3==3.5.4
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0